Add files via upload
This repository is mainly concerned with describing, modelling and estimating stochastic trend model for time series data available in BloaterLH dataset of FSAdata package. The dataset consists of egg deposition (in millions) of age-3 Lake Huron Bloaters (Coregonus hoyi) between years 1981 and 1996. To model this series, we used an integrated autoregressive moving average model (ARIMA) and their properties are thoroughly explored. With a few iterations of the model building strategy, we arrived at the best possible models for the series. Further, various residual plots and error terms for constant variance, normality and independence were used for model diagnostics and check for model adequacy. At last, the original series is forecasted for the next five years (1997-2001) using the parameter estimated of the best model selected.